requirements.txt
streamlit
tensorflow
keras
pillow
matplotlib
